them

+++ manage/public/js/deleteProduct.js
@@ -10,7 +10,15 @@ function initEvents() {
     $(document).on('click', '#btn-delete', function () {
         try {
             var productid = $(this).attr('productid');
-            deleteProduct(productid);
+            $.dialogUpdate({
+                contents: JSMESSAGE.delete_confirm,
+                callback: function (confirm) {
+                    if (confirm) {
+                        deleteProduct(productid);
+                    }
+                }
+            });
+
         } catch (e) {
             alert('remove row ' + e.message);
         }
@@ -35,17 +43,21 @@ function deleteProduct(product_id) {
                 switch (res['status']) {
                     // Success
                     case '200':
-                        alert("Thanh cong");
-                        location.reload();
+                        $.dialogComplete({
+                            contents: JSMESSAGE.delete_complete,
+                            callback: function () {
+                                location.reload();
+                            }
+                        });
+                        // alert("Thanh cong");
+                        // location.reload();
                         break;
                     // Data Validate
                     case '201':
                         alert("Loi 201");
-
                         break;
                     // SQL + PHP Exception
                     case '202':
-
                         alert("Loi 202");
                         break;
                     default:
